+++ apps/mgmt-ui/package.json
@@ -13,7 +13,7 @@
     "@clerk/nextjs": "^4.23.1",
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
-    "@lekko/node-server-sdk": "^0.0.9",
+    "@lekko/react-sdk": "^0.0.2-beta-5",
     "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.11.12",
     "@mui/x-data-grid": "^6.0.0",
