+++ package.json
@@ -67,8 +67,5 @@
     "yaml": "2.2.2",
     "@sentry/node": "7.64.0",
     "async": "3.2.4"
-  },
-  "dependencies": {
-    "@lekko/react-sdk": "^0.0.2-beta-5"
   }
 }
